lib: move fdn and gui into lib
