+++ src/clover_sketches_lib/lib/fdn/draw_fdn.hpp
@@ -4,6 +4,6 @@
 // Copyright (C) 2025  Rob W. Albus
 // Licensed under the GPLv3. See LICENSE for details.
 
-#include "instruments/fdn.hpp"
+#include "lib/fdn/fdn.hpp"
 
 void fdn_component(const char* id, fdn_8_019* fdn_L, fdn_8_019* fdn_R);
